save_messages/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.1
pyee==11.1.0
python-dotenv==1.0.1
slack-sdk==3.11.0
slackeventsapi==3.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
